+++ addons/binding/org.openhab.binding.lgtvserial/src/main/java/org/openhab/binding/lgtvserial/internal/protocol/serial/commands/BasePercentCommand.java
@@ -10,6 +10,7 @@ package org.openhab.binding.lgtvserial.internal.protocol.serial.commands;
 
 import org.eclipse.smarthome.core.library.types.PercentType;
 import org.openhab.binding.lgtvserial.internal.protocol.serial.LGSerialResponse;
+import org.openhab.binding.lgtvserial.internal.protocol.serial.responses.PercentResponse;
 
 /**
  * This command is the base command to handle percent type commands (0-100) in hex format on the wire.
