+++ addons/binding/org.openhab.binding.lgtvserial/src/main/java/org/openhab/binding/lgtvserial/internal/protocol/serial/responses/PercentResponse.java
@@ -6,20 +6,19 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.lgtvserial.internal.protocol.serial.commands;
+package org.openhab.binding.lgtvserial.internal.protocol.serial.responses;
 
-import org.eclipse.smarthome.core.library.types.OnOffType;
-import org.eclipse.smarthome.core.library.types.StringType;
+import org.eclipse.smarthome.core.library.types.PercentType;
 import org.eclipse.smarthome.core.types.State;
 import org.openhab.binding.lgtvserial.internal.protocol.serial.LGSerialResponse;
 
 /**
- * This class represents an ON/OFF response.
+ * This class represents a percentage response.
  *
  * @author Richard Lavoie - Initial contribution
  *
  */
-public class OnOffResponse implements LGSerialResponse {
+public class PercentResponse implements LGSerialResponse {
 
     private int setId;
 
@@ -27,15 +26,10 @@ public class OnOffResponse implements LGSerialResponse {
 
     private State state;
 
-    public OnOffResponse(int setId, boolean success, String data) {
+    public PercentResponse(int setId, boolean success, PercentType state) {
         this.setId = setId;
         this.success = success;
-
-        if (success) {
-            state = data.equals("01") ? OnOffType.ON : OnOffType.OFF;
-        } else {
-            state = new StringType(data);
-        }
+        this.state = state;
     }
 
     @Override
